+++ apps/api/services/rate_info_service.py
@@ -1,0 +1,9 @@
+from apps.core.models import RateInfo
+
+
+def get_rate_info():
+    """
+    Функция возвращает последний rate info
+    :return: rate info
+    """
+    return RateInfo.objects.last()