+++ apps/api/urls/branch_urls.py
@@ -4,6 +4,6 @@ from ..views import branch_views as views
 
 app_name = 'Branchs endpoints'
 urlpatterns = [
-    path('read/list/<int:pk>', views.read_branch_list, name='read_list'),
-    path('read/<int:pk>', views.read_branch, name='read'),
+    path('read/list/<int:pk>', views.read_branch_list_view, name='read_list'),
+    path('read/<int:pk>', views.read_branch_view, name='read'),
 ]