+++ apps/api/urls/qrcode_urls.py
@@ -4,7 +4,7 @@ from ..views import qrcode_views as views
 
 app_name = 'QRCode endpoints'
 urlpatterns = [
-    path('create/', views.create_qrcode, name='create'),
-    path('read/<int:pk>', views.read_qrcode, name='read'),
-    path('read/all', views.read_all_qrcodes, name='read_all'),
+    path('create/', views.create_qrcode_view, name='create'),
+    path('read/<int:pk>', views.read_qrcode_view, name='read'),
+    path('read/all', views.read_all_qrcodes_view, name='read_all'),
 ]