+++ apps/api/views/user_views.py
@@ -43,6 +43,9 @@ class MyTokenObtainPairView(TokenObtainPairView):
             400: openapi.Response(
                 description='Ошибка авторизации'
             ),
+            401: openapi.Response(
+                description='Пользователь с такими данными не найден'
+            ),
             405: openapi.Response(
                 description='Данный метод запроса запрещен'
             )
@@ -112,7 +115,7 @@ class MyTokenObtainPairView(TokenObtainPairView):
     operation_summary='Зарегистрировать user'
 )
 @api_view(['POST'])
-def register_user(request):
+def register_user_view(request):
     """Контроллер для регистрации новых пользователей"""
     data = request.data
     try:
@@ -151,7 +154,7 @@ def register_user(request):
 )
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def read_user_profile(request):
+def read_user_profile_view(request):
     """Контроллер для отдачи информации о текущем пользователе"""
     user = request.user
     serializer = UserSerializer(user, many=False)
@@ -234,7 +237,7 @@ def read_user_profile(request):
 )
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
-def update_user_profile(request):
+def update_user_profile_view(request):
     """Контроллер обновления пользовательских настроек"""
     user = request.user
     data = request.data