+++ frontend/src/components/Brands.js
@@ -7,77 +7,77 @@ import "../css/brands.css";
 const Brands = () => {
   return(
     <>
-    <div id="tabforicons">
+    <div id="tab-for-icons">
       <div className="brandicon">
-        <div className="iconincircle">
-          <img src="images/main/image49.png"/>
+        <div className="icon-in-circle">
+          <img src="/images/main/image49.png"/>
         </div>
         <h5>Яндекс.Карты</h5>
         <p>55 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image50.png"/>
         </div>
         <h5>Google карты</h5>
         <p>30 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image51.png"/>
         </div>
         <h5>2ГИС</h5>
         <p>30 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image52.png"/>
         </div>
         <h5>Wildberries</h5>
         <p>92 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image54.png"/>
         </div>
         <h5>Zoon</h5>
         <p>25 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image55.png"/>
         </div>
         <h5>OZON</h5>
         <p>80 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image56.png"/>
         </div>
         <h5>Yell</h5>
         <p>3 млн чел./месяц</p>
       </div>
       <div className="brandicon">
-        <div className="iconincircle">
+        <div className="icon-in-circle">
           <img src="/images/main/image57.png"/>
         </div>
         <h5>Flamp</h5>
         <p>50 млн чел./месяц</p>
       </div>
     </div>
-    <div id="sliderforicons">
+    <div id="slider-for-icons">
       <Carousel>
         <Carousel.Item>
           <div className="slide">
             <div className="brandicon">
-              <div className="iconincircle">
-                <img src="images/main/image49.png"/>
+              <div className="icon-in-circle">
+                <img src="/images/main/image49.png"/>
               </div>
               <h5>Яндекс.Карты</h5>
               <p>55 млн чел./месяц</p>
             </div>
             <div className="brandicon">
-              <div className="iconincircle">
+              <div className="icon-in-circle">
                 <img src="/images/main/image50.png"/>
               </div>
               <h5>Google карты</h5>
@@ -88,14 +88,14 @@ const Brands = () => {
         <Carousel.Item>
         <div className="slide">
           <div className="brandicon">
-            <div className="iconincircle">
-              <img src="images/main/image51.png"/>
+            <div className="icon-in-circle">
+              <img src="/images/main/image51.png"/>
             </div>
             <h5>2ГИС</h5>
             <p>30 млн чел./месяц</p>
           </div>
           <div className="brandicon">
-            <div className="iconincircle">
+            <div className="icon-in-circle">
               <img src="/images/main/image52.png"/>
             </div>
             <h5>Wildberries</h5>
@@ -106,14 +106,14 @@ const Brands = () => {
         <Carousel.Item>
         <div className="slide">
           <div className="brandicon">
-            <div className="iconincircle">
-              <img src="images/main/image54.png"/>
+            <div className="icon-in-circle">
+              <img src="/images/main/image54.png"/>
             </div>
             <h5>Zoon</h5>
             <p>24 млн чел./месяц</p>
           </div>
           <div className="brandicon">
-            <div className="iconincircle">
+            <div className="icon-in-circle">
               <img src="/images/main/image55.png"/>
             </div>
             <h5>OZON</h5>
@@ -124,14 +124,14 @@ const Brands = () => {
         <Carousel.Item>
         <div className="slide">
           <div className="brandicon">
-            <div className="iconincircle">
-              <img src="images/main/image54.png"/>
+            <div className="icon-in-circle">
+              <img src="/images/main/image54.png"/>
             </div>
             <h5>Yell</h5>
             <p>3 млн чел./месяц</p>
           </div>
           <div className="brandicon">
-            <div className="iconincircle">
+            <div className="icon-in-circle">
               <img src="/images/main/image55.png"/>
             </div>
             <h5>Flamp</h5>