+++ CHANGELOG.md
@@ -2,18 +2,41 @@
 
 ## [Unreleased]
 
+## [0.0.13] - 2022-10-21
+
+### Добавлено 
+
+- Docs for new service [@fringer2423](https://github.com/fringer2423)
+
+### Изменено
+
+- Context fix urls [@fringer2423](https://github.com/fringer2423)
+- Refactor code
+
 ## [0.0.12] - 2022-10-20
 
 ### Добавлено
 
 - Компонент ListOfRec и css [AlexTrubkina](https://github.com/AlexTrubkina)
-- CSS для кнопок [AlexTrubkina](https://github.com/AlexTrubkina)
+- CSS для кнопок 
+- Модели Rate и RateInfo [@fringer2423](https://github.com/fringer2423)
+- Обработка 401 статуса в авторизации
+- Rate info serializer
+- Rate info endpoints
+- Rate info service
+- Rate serializer
+- Rate endpoints
+- Rate service
 
 ### Изменено
 
+- Connect fix [@fringer2423](https://github.com/fringer2423)
+- User model
+- Добавлено поле rate в UserSerializer
 - Компонент Prices.js (добавлена карусель) [AlexTrubkina](https://github.com/AlexTrubkina)
-- Компоненты Prices и HomeScreen [AlexTrubkina](https://github.com/AlexTrubkina)
-- CSS buttons, main, purplecircle [AlexTrubkina](https://github.com/AlexTrubkina)
+- Компоненты Prices и HomeScreen 
+- CSS buttons, main, purplecircle 
+
 
 ## [0.0.11] - 2022-10-19